# 450/openapi_export.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "电商系统API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/orders": {
      "post": {
        "summary": "创建订单",
        "description": "提交新订单",
        "tags": [
          "订单"
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "required": true,
            "description": "产品ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "description": "购买数量",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "coupon_code",
            "in": "query",
            "required": false,
            "description": "优惠码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address_id",
            "in": "query",
            "required": true,
            "description": "收货地址ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应"
          },
          "400": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "error": "product_not_found",
                  "message": "产品不存在"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "summary": "获取用户列表",
        "description": "根据筛选条件分页查询用户信息",
        "tags": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "页码，默认1",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "每页数量，默认20",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "description": "搜索关键词",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": true,
            "description": "角色筛选 (admin/user/guest)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应"
          }
        }
      },
      "post": {
        "summary": "创建新用户",
        "description": "提交用户信息创建新账号",
        "tags": [],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "description": "用户名，2-50个字符",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "description": "邮箱地址",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "description": "密码，至少8位",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "description": "角色，默认user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "https://api.example.com"
    }
  ],
...
}